# Add support for memoized methods with block arguments