See #46.

Two calls to a memoized method with different block arguments will now be cached separately.