@flash-gordon 2021-01-26
I think this is not practical. Every call of a memoized method normally creates a new block with a new hash value. The only way not to have a leak is to ensure only a finite amount of calls done or blocks are cached somewhere else. I don't think this API is worth adding.