@oleander 2021-01-26
@flash-gordon 

> Every call of a memoized method normally creates a new block with a new hash value. 

``` ruby
def test(&block)
  puts block.hash
end

2.times do
  test do
    # NOP
  end
end
```
 Yields:

```
4375437158601092624
4375437158601092624
```

This one outputs the same hash value for both calls to test tho, or maybe I misunderstood what you meant. If the block's hash isn't part of the cache key then we should at least pass the block to super or raise an exception notifying the caller that blocks aren't supported ツ