@flash-gordon 2021-01-26
@oleander yeah, you're right. I wonder where my prev understanding is coming from... Anyway, this makes it practical for me :) Although, it's not possible to determine if a method accepts a block because it can use the `yield` keyword.