@flash-gordon 2021-01-26
I find this odd though I do understand how it works
```ruby
def capture(&block)
  block
end

def wrap(value)
  capture { value }
end

f = wrap(1)
g = wrap(2)

puts(f.hash == g.hash) # => true
puts(f.() == g.())     # => false
```