@oleander 2021-01-26
The new implantation doesn't have to be perfect tho, I would argue—just better than the current one. If we can support explicitly declared blocks I would call it a win. Under the assumption that the change is backward compatible with the current version. 

We could check for blocks using "block_given?" And check that a block is included in "method.parameters.to_h.key?(:block)". If it isn't then "yield" is used and we should fall back to the previous implementation, not cache the call or raise an error.