@solnic 2021-01-27
> @oleander yeah, you're right. I wonder where my prev understanding is coming from

@flash-gordon actually, my initial gut-feel reaction whenever I see blocks and memoization is to panic, so I totally understand your initial concern! It *seems* like we're good here though

@oleander @flash-gordon I think it's worth adding and I also think it's worth having a generic API for redefining methods while preserving signatures. We use all sorts of method-overriding and method decoration in dry-rb and rom-rb ecosystems so this would be beneficial (assuming it's possible to have a stable implementation 🙂).

Also, to make sure we're on the same page:

> Ah, I understand what you mean now. You want the memorized method to not only behave in the same way as the original one but also to return the same stats as the original, i.e trough method(:mymethod).parameters.

This is exactly what I meant, sorry if I rushed with my initial explanation! I believe this would avoid any potential confusion with overridden methods and also open up doors for some new possibilities where you may want to use parameters info to do...well, something cool 🤓 

So, we could use it for memoization API here and if it works well, we could promote it to a 1st-class API that could be used in other gems.