@oleander 2021-01-27
Ran some benchmarks against HEAD using this implementation based on `module_eval` (Ruby 2.5.1).

``` ruby
require "bundler/setup"
require "benchmark/ips"

Bundler.require

require 'dry/core/memoizable'

class Test
  include Dry::Core::Memoizable

  def call(*args, **kwargs)
    # NOP
  end

  memoize :call
end

test = Test.new
block = -> { }

Benchmark.ips do |x|
  x.report("new") { test.call(:arg, key: :value) }
end
```

## Without block

### HEAD

```
Warming up --------------------------------------
                 new     3.627k i/100ms
Calculating -------------------------------------
                 new     37.082k (± 4.7%) i/s -    184.977k in   4.999907s
```

### module_eval

```
Warming up --------------------------------------
                 new     3.636k i/100ms
Calculating -------------------------------------
                 new     35.813k (± 6.8%) i/s -    178.164k in   5.000445s
```

## With block

### HEAD

```
Warming up --------------------------------------
                 new     2.876k i/100ms
Calculating -------------------------------------
                 new     28.488k (± 7.1%) i/s -    143.800k in   5.077245s
```

### module_eval

```
Warming up --------------------------------------
                 new     2.854k i/100ms
Calculating -------------------------------------
                 new     28.937k (± 4.3%) i/s -    145.554k in   5.039862s
```