@oleander 2021-01-27
Rebased with master to add support for BasicObject.