@oleander 2021-01-27
> we could promote it to a 1st-class API that could be used in other gems.

I didn't know that wasn't already the case. I've been using it outside the dry-suite for months :)