@solnic 2021-01-28
>> we could promote it to a 1st-class API that could be used in other gems.
> I didn't know that wasn't already the case. I've been using it outside the dry-suite for months :)

@oleander sorry, I meant the API for redefining methods while preserving original sigs! Memoizable is a 1st-class, public API. We should have it documented one day 😓 I want to wrap up dry-core 1.0.0 *soon* so this will be a good moment to revamp the docs.