@flash-gordon 2021-06-12
As I'm adding specs for this it turned out `Proc#hash` can return different values for the same block:
```ruby
def capture(&block)
  block.hash
end

ids = Array.new(1000) { capture { :foo } }.uniq

puts(ids.size) # this returns different values on each call
```
This leads to a memory leak. I didn't look into details yet but my gut feelings tell me we won't be able to "fix" this. In this case, I would suggest deprecating using blocks for memorization. It's still fine to pass them, it's just `Proc#hash` is unreliable for our purposes.