@flash-gordon 2021-06-13
will wait for comments on https://bugs.ruby-lang.org/issues/17951